backend/requirements.txt
alembic==1.16.5
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==22.0.0
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
msgspec==0.19.0
playwright==1.54.0
pyee==13.0.0
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.43
stripe==12.5.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
